cic(#1513): lift the send pipeline out of the store module
`compose.ts` imports `commands/*`, so anything a command handler needs
cannot stay in `compose.ts` without closing an import cycle — and the
send pipeline is exactly what the six arms #1513 names still reach for.
This moves the cluster to `lib/sendPipeline.ts`: `draftLines`,
`wireBody`, the three #666 retry constants, `sleep`, `isSendThrottled`,
`retryAfterMs`, `PacedSend`, `planSends`, `drainPaced`, `sendBodyLines`
and `sendFanOut`. `compose.ts` goes 1747 -> 1557 lines.

It moves WHOLE rather than in pieces, and that was measured before it
was asserted: every one of those symbols was module-scope (the
`identityScopedStore` closure only opens at the old :434), their only
outward dependencies are two PURE leaves — `messageLines.ts` and
`ctcpAction.ts`, zero imports each — plus `scrollback` and `ApiError`,
and none of the cluster internals is referenced anywhere outside
`compose.ts`. The precedent is written in the code it depends on:
`ctcpAction.ts:23-26` records #1192 taking `ctcpFrame` out of
`compose.ts` for this same reason, "dispatch without importing
`compose`".

`draftLines` and `wireBody` travel with the cluster instead of to a leaf
of their own because all three of their call sites use them as a PAIR.
`draftFramePreview` does NOT travel: it needs `parseSlash` and
`aliases()`, which are composer concerns, so it stays and now reads the
two helpers across the new boundary — which is precisely the edge the
mutant below exercises.

`FANOUT_CONFIRM_THRESHOLD` also stays. Its only call site is the
`ame`/`amsg` arm, not `sendFanOut`, so it belongs with the arm whenever
that moves — the same rule that took `listModeQueryLetter` into
`mode.ts` in #1396 slice 2a.

`sendBodyLines` keeps its two component callers, repointed at the new
module together with their `vi.mock` targets under `ServiceModal.test.tsx`
and `RegistrationWizardModal.test.tsx` — the relocation #1396 slice 2b
predicted when it refused to move `service-modal`.

Proven, not assumed. Suite green before and after the move at the same
290 files / 5615 tests. The mutant is the one the plan committed to: drop
the CTCP ACTION framing in `wireBody`, and require the SAME SET of reds
before and after the move, not merely the same count. Measured on both
sides: seven, identical, and split across two families — one PREVIEW
("charges /me its CTCP ACTION envelope, like the send path does"), five
WIRE (/me, multiline /me, the #723 partial, the #1126 scrub, /ame), one
characterization net. The preview red is the load-bearing one: had it
gone green while the wire reds stayed, the helper would have been
DUPLICATED rather than moved, and no plain green would have said so.
Mutant restored byte-identical (`cmp`) on both sides.
